refactor: Auth API
